REF: test_fillna
